ai-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
anthropic==0.40.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.28.0
redis==5.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
structlog==24.4.0
